+++ ReleaseNotes.md
@@ -1,8 +1,8 @@
-# 2020.2.xx.xxx-pre-release
+# 2020.2.15.xxx-pre-release
 F# niggles part 2
 * Explicitly exempt the `get_` and `set_` prefixes of getter and setter methods from the `AvoidNonAlphanumericIdentifierRule` since that was not already a thing.
 * Make `AvoidUnneededUnboxingRule` not applicable to `[CompilerGenerated]` functions (e.g. Union case `CompareTo`)
-* Exempt debugger-related generated types related to union types from `AvoidUnsealedUninheritedInternalTypeRule`
+* Exempt debugger-related generated types related to union types from `AvoidUnsealedUninheritedInternalTypeRule` and `UseCorrectCasingRule`
 * Exempt getter and setter methods from `ConsiderConvertingMethodToPropertyRule` (well, duh!)
 * Exempt types with only fully `[CompilerGenerated]` `Equals` and `CompareTo` methods from `ImplementIComparableCorrectlyRule`; also explicitly exempt record types